@@ -1,11 +1,13 @@
 {
   "name": "ts-gsm-modem",
-  "version": "0.19.0",
+  "version": "0.19.1",
   "description": "GSM Modem manager",
   "main": "./dist/lib/index.js",
   "types": "./dist/lib/index.d.ts",
   "scripts": {
-    "tsc": "./node_modules/.bin/tsc -p ./tsconfig.json",
+    "postinstall": "npm run remove-bold-color",
+    "remove-bold-color": "sed -i '/bold: string;/d' ./node_modules/@types/colors/index.d.ts",
+    "tsc": "node ./node_modules/typescript/bin/tsc -p ./tsconfig.json",
     "tsc-w": "node ./node_modules/typescript/bin/tsc -w -p ./tsconfig.json",
     "test": "DEBUG=_* node ./dist/test/main"
   },
@@ -24,25 +26,25 @@
   "author": "Joseph Garrone",
   "license": "SEE LICENSE IN LICENSE",
   "dependencies": {
-    "at-messages-parser": "git+https://a30f7432b117d8320cc21cc03513817c3f08d4b6:x-oauth-basic@github.com/garronej/at-messages-parser.git",
+    "at-messages-parser": "git+https://206e4c24cb9a1e4a8d42b7ac86e4a026b02cee42:x-oauth-basic@github.com/garronej/at-messages-parser.git",
     "colors": "^1.1.2",
     "debug": "^2.6.1",
     "legacy-encoding": "^3.0.0",
-    "node-python-messaging": "git+https://a30f7432b117d8320cc21cc03513817c3f08d4b6:x-oauth-basic@github.com/garronej/node-python-messaging.git",
+    "node-python-messaging": "git+https://206e4c24cb9a1e4a8d42b7ac86e4a026b02cee42:x-oauth-basic@github.com/garronej/node-python-messaging.git",
     "rejection-tracker": "garronej/rejection-tracker",
     "run-exclusive": "github:garronej/run-exclusive",
     "serialport": "EmergingTechnologyAdvisors/node-serialport#4.0.7",
     "timer-extended": "garronej/timer-extended",
     "trackable-map": "github:garronej/trackable-map",
     "ts-events-extended": "garronej/ts-events-extended",
-    "gsm-modem-connection": "git+https://a30f7432b117d8320cc21cc03513817c3f08d4b6:x-oauth-basic@github.com/garronej/gsm-modem-connection"
+    "gsm-modem-connection": "git+https://206e4c24cb9a1e4a8d42b7ac86e4a026b02cee42:x-oauth-basic@github.com/garronej/gsm-modem-connection"
   },
   "devDependencies": {
-    "gsm-modem-connection": "git+https://a30f7432b117d8320cc21cc03513817c3f08d4b6:x-oauth-basic@github.com/garronej/gsm-modem-connection.git",
-    "typescript": "^2.2.1",
-    "@types/colors": "^1.1.1",
+    "gsm-modem-connection": "git+https://206e4c24cb9a1e4a8d42b7ac86e4a026b02cee42:x-oauth-basic@github.com/garronej/gsm-modem-connection.git",
+    "typescript": "^2.7.1",
+    "@types/colors": "^1.1.3",
     "@types/chai": "^3.4.34",
     "@types/debug": "0.0.29",
-    "@types/node": "^4.2.2"
+    "@types/node": "^8.9.1"
   }
 }
